Calculating Expected Return, Beta and CAPM
CAPM is a finance model that establishes a linear relationship between the required return on an investment and risk. The model is based on the relationship between an asset's beta, the risk-free rate (typically the Treasury bill rate), and the equity risk premium, or the expected return on the market minus the risk-free rate.